Move NameContainsKeywordsPredicate from person to deck package
