+++ src/main/java/seedu/address/model/deck/NameContainsKeywordsPredicate.java
@@ -1,10 +1,9 @@
-package seedu.address.model.person;
+package seedu.address.model.deck;
 
 import java.util.List;
 import java.util.function.Predicate;
 
 import seedu.address.commons.util.StringUtil;
-import seedu.address.model.deck.Card;
 
 /**
  * Tests that a {@code Card}'s {@code Name} matches any of the keywords given.
