+++ src/test/java/seedu/address/logic/commands/ClearCommandTest.java
@@ -1,14 +1,14 @@
 package seedu.address.logic.commands;
 
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
+import static seedu.address.testutil.TypicalCards.getTypicalTopDeck;
 
 import org.junit.Test;
 
 import seedu.address.logic.CommandHistory;
-import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
+import seedu.address.model.TopDeck;
 import seedu.address.model.UserPrefs;
 
 public class ClearCommandTest {
@@ -26,9 +26,9 @@ public class ClearCommandTest {
 
     @Test
     public void execute_nonEmptyAddressBook_success() {
-        Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
-        Model expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs());
-        expectedModel.setTopDeck(new AddressBook());
+        Model model = new ModelManager(getTypicalTopDeck(), new UserPrefs());
+        Model expectedModel = new ModelManager(getTypicalTopDeck(), new UserPrefs());
+        expectedModel.setTopDeck(new TopDeck());
         expectedModel.commitTopDeck();
 
         assertCommandSuccess(new ClearCommand(), model, commandHistory, ClearCommand.MESSAGE_SUCCESS, expectedModel);
