+++ src/test/java/seedu/address/logic/parser/TopDeckParserTest.java
@@ -27,7 +27,7 @@ import seedu.address.logic.commands.RedoCommand;
 import seedu.address.logic.commands.SelectCommand;
 import seedu.address.logic.commands.UndoCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.model.deck.NameContainsKeywordsPredicate;
 import seedu.address.model.person.Person;
 import seedu.address.testutil.EditCardDescriptorBuilder;
 import seedu.address.testutil.PersonBuilder;
